+++ src/main/java/nextstep/app/infrastructure/GithubClient.java
@@ -3,9 +3,9 @@ package nextstep.app.infrastructure;
 import nextstep.app.infrastructure.dto.AccessRequest;
 import nextstep.app.infrastructure.dto.AccessResponse;
 import nextstep.app.infrastructure.dto.UserResponse;
-import nextstep.security.authentication.BaseOauth2User;
-import nextstep.security.authentication.OAuth2User;
-import nextstep.security.authentication.Oauth2UserService;
+import nextstep.security.oauth2user.BaseOauth2User;
+import nextstep.security.oauth2user.OAuth2User;
+import nextstep.security.oauth2user.Oauth2UserService;
 import nextstep.security.exception.AuthenticationException;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Component;
